move evaluation of expressions and statements to separate files
