+++ runtime/interpreter/interpreter.go
@@ -21,9 +21,7 @@ package interpreter
 import (
 	"fmt"
 	goRuntime "runtime"
-	"sort"
 
-	"github.com/onflow/cadence/fixedpoint"
 	"github.com/onflow/cadence/runtime/activations"
 	"github.com/onflow/cadence/runtime/ast"
 	"github.com/onflow/cadence/runtime/common"
@@ -975,32 +973,6 @@ func (interpreter *Interpreter) VisitBlock(block *ast.Block) ast.Repr {
 		})
 }
 
-func (interpreter *Interpreter) visitStatements(statements []ast.Statement) Trampoline {
-	count := len(statements)
-
-	// no statements? stop
-	if count == 0 {
-		// NOTE: no result, so it does *not* act like a return-statement
-		return Done{}
-	}
-
-	statement := statements[0]
-
-	// interpret the first statement, then the remaining ones
-	return StatementTrampoline{
-		F: func() Trampoline {
-			return statement.Accept(interpreter).(Trampoline)
-		},
-		Interpreter: interpreter,
-		Statement:   statement,
-	}.FlatMap(func(returnValue interface{}) Trampoline {
-		if _, isReturn := returnValue.(controlReturn); isReturn {
-			return Done{Result: returnValue}
-		}
-		return interpreter.visitStatements(statements[1:])
-	})
-}
-
 func (interpreter *Interpreter) VisitFunctionBlock(_ *ast.FunctionBlock) ast.Repr {
 	// NOTE: see visitBlock
 	panic(errors.NewUnreachableError())
@@ -1098,284 +1070,6 @@ func (interpreter *Interpreter) VisitCondition(condition *ast.Condition) ast.Rep
 	return Done{Result: result}
 }
 
-func (interpreter *Interpreter) VisitReturnStatement(statement *ast.ReturnStatement) ast.Repr {
-	// NOTE: returning result
-
-	var value Value
-	if statement.Expression == nil {
-		value = VoidValue{}
-	} else {
-		value = interpreter.evalExpression(statement.Expression)
-
-		valueType := interpreter.Program.Elaboration.ReturnStatementValueTypes[statement]
-		returnType := interpreter.Program.Elaboration.ReturnStatementReturnTypes[statement]
-
-		// NOTE: copy on return
-		value = interpreter.copyAndConvert(value, valueType, returnType)
-	}
-	return Done{Result: functionReturn{value}}
-}
-
-func (interpreter *Interpreter) VisitBreakStatement(_ *ast.BreakStatement) ast.Repr {
-	return Done{Result: controlBreak{}}
-}
-
-func (interpreter *Interpreter) VisitContinueStatement(_ *ast.ContinueStatement) ast.Repr {
-	return Done{Result: controlContinue{}}
-}
-
-func (interpreter *Interpreter) VisitIfStatement(statement *ast.IfStatement) ast.Repr {
-	switch test := statement.Test.(type) {
-	case ast.Expression:
-		return interpreter.visitIfStatementWithTestExpression(test, statement.Then, statement.Else)
-	case *ast.VariableDeclaration:
-		return interpreter.visitIfStatementWithVariableDeclaration(test, statement.Then, statement.Else)
-	default:
-		panic(errors.NewUnreachableError())
-	}
-}
-
-func (interpreter *Interpreter) visitIfStatementWithTestExpression(
-	test ast.Expression,
-	thenBlock, elseBlock *ast.Block,
-) Trampoline {
-
-	value := interpreter.evalExpression(test).(BoolValue)
-	if value {
-		return thenBlock.Accept(interpreter).(Trampoline)
-	} else if elseBlock != nil {
-		return elseBlock.Accept(interpreter).(Trampoline)
-	}
-
-	// NOTE: no result, so it does *not* act like a return-statement
-	return Done{}
-}
-
-func (interpreter *Interpreter) visitIfStatementWithVariableDeclaration(
-	declaration *ast.VariableDeclaration,
-	thenBlock, elseBlock *ast.Block,
-) Trampoline {
-
-	result := interpreter.evalExpression(declaration.Value)
-
-	if someValue, ok := result.(*SomeValue); ok {
-
-		targetType := interpreter.Program.Elaboration.VariableDeclarationTargetTypes[declaration]
-		valueType := interpreter.Program.Elaboration.VariableDeclarationValueTypes[declaration]
-		unwrappedValueCopy := interpreter.copyAndConvert(someValue.Value, valueType, targetType)
-
-		interpreter.activations.PushNewWithCurrent()
-		interpreter.declareVariable(
-			declaration.Identifier.Identifier,
-			unwrappedValueCopy,
-		)
-
-		return thenBlock.Accept(interpreter).(Trampoline).
-			Then(func(_ interface{}) {
-				interpreter.activations.Pop()
-			})
-	} else if elseBlock != nil {
-		return elseBlock.Accept(interpreter).(Trampoline)
-	}
-
-	// NOTE: ignore result, so it does *not* act like a return-statement
-	return Done{}
-}
-
-func (interpreter *Interpreter) VisitSwitchStatement(switchStatement *ast.SwitchStatement) ast.Repr {
-
-	var visitCase func(i int, testValue EquatableValue) Trampoline
-	visitCase = func(i int, testValue EquatableValue) Trampoline {
-
-		// If no cases are left to evaluate, return (base case)
-
-		if i >= len(switchStatement.Cases) {
-			// NOTE: no result, so it does *not* act like a return-statement
-			return Done{}
-		}
-
-		switchCase := switchStatement.Cases[i]
-
-		runStatements := func() Trampoline {
-			// NOTE: the new block ensures that a new block is introduced
-
-			block := &ast.Block{
-				Statements: switchCase.Statements,
-			}
-
-			return block.Accept(interpreter).(Trampoline).
-				FlatMap(func(value interface{}) Trampoline {
-
-					if _, ok := value.(controlBreak); ok {
-						return Done{}
-					}
-
-					return Done{Result: value}
-				})
-		}
-
-		// If the case has no expression it is the default case.
-		// Evaluate it, i.e. all statements
-
-		if switchCase.Expression == nil {
-			return runStatements()
-		}
-
-		// The case has an expression.
-		// Evaluate it and compare it to the test value
-
-		result := interpreter.evalExpression(switchCase.Expression)
-
-		caseValue := result.(EquatableValue)
-
-		// If the test value and case values are equal,
-		// evaluate the case's statements
-
-		if testValue.Equal(interpreter, caseValue) {
-			return runStatements()
-		}
-
-		// If the test value and the case values are unequal,
-		// try the next case (recurse)
-
-		return visitCase(i+1, testValue)
-	}
-
-	testValue := interpreter.evalExpression(switchStatement.Expression).(EquatableValue)
-	return visitCase(0, testValue)
-}
-
-func (interpreter *Interpreter) VisitWhileStatement(statement *ast.WhileStatement) ast.Repr {
-
-	value := interpreter.evalExpression(statement.Test).(BoolValue)
-	if !value {
-		return Done{}
-	}
-
-	interpreter.reportLoopIteration(statement)
-
-	return statement.Block.Accept(interpreter).(Trampoline).
-		FlatMap(func(value interface{}) Trampoline {
-
-			switch value.(type) {
-			case controlBreak:
-				return Done{}
-
-			case controlContinue:
-				// NO-OP
-
-			case functionReturn:
-				return Done{Result: value}
-			}
-
-			// recurse
-			return statement.Accept(interpreter).(Trampoline)
-		})
-}
-
-func (interpreter *Interpreter) VisitForStatement(statement *ast.ForStatement) ast.Repr {
-	interpreter.activations.PushNewWithCurrent()
-
-	variable := interpreter.declareVariable(
-		statement.Identifier.Identifier,
-		nil,
-	)
-
-	var loop func(i, count int, values []Value) Trampoline
-	loop = func(i, count int, values []Value) Trampoline {
-
-		if i == count {
-			return Done{}
-		}
-
-		interpreter.reportLoopIteration(statement)
-
-		variable.Value = values[i]
-
-		return statement.Block.Accept(interpreter).(Trampoline).
-			FlatMap(func(value interface{}) Trampoline {
-
-				switch value.(type) {
-				case controlBreak:
-					return Done{}
-
-				case controlContinue:
-					// NO-OP
-
-				case functionReturn:
-					return Done{Result: value}
-				}
-
-				// recurse
-				if i == count {
-					return Done{}
-				}
-				return loop(i+1, count, values)
-			})
-	}
-
-	values := interpreter.evalExpression(statement.Value).(*ArrayValue).Values[:]
-	count := len(values)
-
-	return loop(0, count, values).
-		Then(func(_ interface{}) {
-			interpreter.activations.Pop()
-		})
-}
-
-func (interpreter *Interpreter) visitPotentialStorageRemoval(expression ast.Expression) Value {
-	movingStorageIndexExpression := interpreter.movingStorageIndexExpression(expression)
-	if movingStorageIndexExpression == nil {
-		return interpreter.evalExpression(expression)
-	}
-
-	getterSetter := interpreter.indexExpressionGetterSetter(movingStorageIndexExpression)
-	value := getterSetter.get()
-	getterSetter.set(NilValue{})
-	return value
-}
-
-// VisitVariableDeclaration first visits the declaration's value,
-// then declares the variable with the name bound to the value
-func (interpreter *Interpreter) VisitVariableDeclaration(declaration *ast.VariableDeclaration) ast.Repr {
-
-	targetType := interpreter.Program.Elaboration.VariableDeclarationTargetTypes[declaration]
-	valueType := interpreter.Program.Elaboration.VariableDeclarationValueTypes[declaration]
-	secondValueType := interpreter.Program.Elaboration.VariableDeclarationSecondValueTypes[declaration]
-
-	result := interpreter.visitPotentialStorageRemoval(declaration.Value)
-
-	valueCopy := interpreter.copyAndConvert(result, valueType, targetType)
-
-	interpreter.declareVariable(
-		declaration.Identifier.Identifier,
-		valueCopy,
-	)
-
-	if declaration.SecondValue == nil {
-		// NOTE: ignore result, so it does *not* act like a return-statement
-		return Done{}
-	}
-
-	return interpreter.visitAssignment(
-		declaration.Transfer.Operation,
-		declaration.Value,
-		valueType,
-		declaration.SecondValue,
-		secondValueType,
-		declaration,
-	)
-}
-
-func (interpreter *Interpreter) movingStorageIndexExpression(expression ast.Expression) *ast.IndexExpression {
-	indexExpression, ok := expression.(*ast.IndexExpression)
-	if !ok || !interpreter.Program.Elaboration.IsResourceMovingStorageIndexExpression[indexExpression] {
-		return nil
-	}
-
-	return indexExpression
-}
-
 func (interpreter *Interpreter) declareValue(declaration ValueDeclaration) *Variable {
 
 	if !declaration.ValueDeclarationAvailable(interpreter.Location) {
@@ -1394,21 +1088,6 @@ func (interpreter *Interpreter) declareVariable(identifier string, value Value) 
 	variable := NewVariable(value)
 	interpreter.setVariable(identifier, variable)
 	return variable
-}
-
-func (interpreter *Interpreter) VisitAssignmentStatement(assignment *ast.AssignmentStatement) ast.Repr {
-	targetType := interpreter.Program.Elaboration.AssignmentStatementTargetTypes[assignment]
-	valueType := interpreter.Program.Elaboration.AssignmentStatementValueTypes[assignment]
-
-	target := assignment.Target
-	value := assignment.Value
-
-	return interpreter.visitAssignment(
-		assignment.Transfer.Operation,
-		target, targetType,
-		value, valueType,
-		assignment,
-	)
 }
 
 func (interpreter *Interpreter) visitAssignment(
@@ -1443,684 +1122,6 @@ func (interpreter *Interpreter) visitAssignment(
 
 	// NOTE: no result, so it does *not* act like a return-statement
 	return Done{}
-}
-
-func (interpreter *Interpreter) VisitSwapStatement(swap *ast.SwapStatement) ast.Repr {
-
-	leftType := interpreter.Program.Elaboration.SwapStatementLeftTypes[swap]
-	rightType := interpreter.Program.Elaboration.SwapStatementRightTypes[swap]
-
-	// Evaluate the left expression
-	leftGetterSetter := interpreter.assignmentGetterSetter(swap.Left)
-	leftValue := leftGetterSetter.get()
-	if interpreter.movingStorageIndexExpression(swap.Left) != nil {
-		leftGetterSetter.set(NilValue{})
-	}
-
-	// Evaluate the right expression
-	rightGetterSetter := interpreter.assignmentGetterSetter(swap.Right)
-	rightValue := rightGetterSetter.get()
-	if interpreter.movingStorageIndexExpression(swap.Right) != nil {
-		rightGetterSetter.set(NilValue{})
-	}
-
-	// Add right value to left target
-	// and left value to right target
-
-	rightValueCopy := interpreter.copyAndConvert(rightValue.(Value), rightType, leftType)
-	leftValueCopy := interpreter.copyAndConvert(leftValue.(Value), leftType, rightType)
-
-	leftGetterSetter.set(rightValueCopy)
-	rightGetterSetter.set(leftValueCopy)
-
-	return Done{}
-}
-
-// assignmentGetterSetter returns a getter/setter function pair
-// for the target expression, wrapped in a trampoline
-//
-func (interpreter *Interpreter) assignmentGetterSetter(target ast.Expression) getterSetter {
-	switch target := target.(type) {
-	case *ast.IdentifierExpression:
-		return interpreter.identifierExpressionGetterSetter(target)
-
-	case *ast.IndexExpression:
-		return interpreter.indexExpressionGetterSetter(target)
-
-	case *ast.MemberExpression:
-		return interpreter.memberExpressionGetterSetter(target)
-	}
-
-	panic(errors.NewUnreachableError())
-}
-
-// identifierExpressionGetterSetter returns a getter/setter function pair
-// for the target identifier expression, wrapped in a trampoline
-//
-func (interpreter *Interpreter) identifierExpressionGetterSetter(identifierExpression *ast.IdentifierExpression) getterSetter {
-	variable := interpreter.findVariable(identifierExpression.Identifier.Identifier)
-	return getterSetter{
-		get: func() Value {
-			return variable.Value
-		},
-		set: func(value Value) {
-			variable.Value = value
-		},
-	}
-}
-
-// indexExpressionGetterSetter returns a getter/setter function pair
-// for the target index expression, wrapped in a trampoline
-//
-func (interpreter *Interpreter) indexExpressionGetterSetter(indexExpression *ast.IndexExpression) getterSetter {
-	typedResult := interpreter.evalExpression(indexExpression.TargetExpression).(ValueIndexableValue)
-	indexingValue := interpreter.evalExpression(indexExpression.IndexingExpression)
-	locationRange := interpreter.locationRange(indexExpression)
-	return getterSetter{
-		get: func() Value {
-			return typedResult.Get(interpreter, locationRange, indexingValue)
-		},
-		set: func(value Value) {
-			typedResult.Set(interpreter, locationRange, indexingValue, value)
-		},
-	}
-}
-
-// memberExpressionGetterSetter returns a getter/setter function pair
-// for the target member expression, wrapped in a trampoline
-//
-func (interpreter *Interpreter) memberExpressionGetterSetter(memberExpression *ast.MemberExpression) getterSetter {
-	target := interpreter.evalExpression(memberExpression.Expression)
-	locationRange := interpreter.locationRange(memberExpression)
-	identifier := memberExpression.Identifier.Identifier
-	return getterSetter{
-		get: func() Value {
-			return interpreter.getMember(target, locationRange, identifier)
-		},
-		set: func(value Value) {
-			interpreter.setMember(target, locationRange, identifier, value)
-		},
-	}
-}
-
-func (interpreter *Interpreter) VisitIdentifierExpression(expression *ast.IdentifierExpression) ast.Repr {
-	name := expression.Identifier.Identifier
-	variable := interpreter.findVariable(name)
-	return variable.Value
-}
-
-func (interpreter *Interpreter) evalExpression(expression ast.Expression) Value {
-	return expression.Accept(interpreter).(Value)
-}
-
-func (interpreter *Interpreter) VisitBinaryExpression(expression *ast.BinaryExpression) ast.Repr {
-	switch expression.Operation {
-	case ast.OperationPlus:
-		left := interpreter.evalExpression(expression.Left).(NumberValue)
-		right := interpreter.evalExpression(expression.Right).(NumberValue)
-		return left.Plus(right)
-
-	case ast.OperationMinus:
-		left := interpreter.evalExpression(expression.Left).(NumberValue)
-		right := interpreter.evalExpression(expression.Right).(NumberValue)
-		return left.Minus(right)
-
-	case ast.OperationMod:
-		left := interpreter.evalExpression(expression.Left).(NumberValue)
-		right := interpreter.evalExpression(expression.Right).(NumberValue)
-		return left.Mod(right)
-
-	case ast.OperationMul:
-		left := interpreter.evalExpression(expression.Left).(NumberValue)
-		right := interpreter.evalExpression(expression.Right).(NumberValue)
-		return left.Mul(right)
-
-	case ast.OperationDiv:
-		left := interpreter.evalExpression(expression.Left).(NumberValue)
-		right := interpreter.evalExpression(expression.Right).(NumberValue)
-		return left.Div(right)
-
-	case ast.OperationBitwiseOr:
-		left := interpreter.evalExpression(expression.Left).(IntegerValue)
-		right := interpreter.evalExpression(expression.Right).(IntegerValue)
-		return left.BitwiseOr(right)
-
-	case ast.OperationBitwiseXor:
-		left := interpreter.evalExpression(expression.Left).(IntegerValue)
-		right := interpreter.evalExpression(expression.Right).(IntegerValue)
-		return left.BitwiseXor(right)
-
-	case ast.OperationBitwiseAnd:
-		left := interpreter.evalExpression(expression.Left).(IntegerValue)
-		right := interpreter.evalExpression(expression.Right).(IntegerValue)
-		return left.BitwiseAnd(right)
-
-	case ast.OperationBitwiseLeftShift:
-		left := interpreter.evalExpression(expression.Left).(IntegerValue)
-		right := interpreter.evalExpression(expression.Right).(IntegerValue)
-		return left.BitwiseLeftShift(right)
-
-	case ast.OperationBitwiseRightShift:
-		left := interpreter.evalExpression(expression.Left).(IntegerValue)
-		right := interpreter.evalExpression(expression.Right).(IntegerValue)
-		return left.BitwiseRightShift(right)
-
-	case ast.OperationLess:
-		left := interpreter.evalExpression(expression.Left).(NumberValue)
-		right := interpreter.evalExpression(expression.Right).(NumberValue)
-		return left.Less(right)
-
-	case ast.OperationLessEqual:
-		left := interpreter.evalExpression(expression.Left).(NumberValue)
-		right := interpreter.evalExpression(expression.Right).(NumberValue)
-		return left.LessEqual(right)
-
-	case ast.OperationGreater:
-		left := interpreter.evalExpression(expression.Left).(NumberValue)
-		right := interpreter.evalExpression(expression.Right).(NumberValue)
-		return left.Greater(right)
-
-	case ast.OperationGreaterEqual:
-		left := interpreter.evalExpression(expression.Left).(NumberValue)
-		right := interpreter.evalExpression(expression.Right).(NumberValue)
-		return left.GreaterEqual(right)
-
-	case ast.OperationEqual:
-		left := interpreter.evalExpression(expression.Left)
-		right := interpreter.evalExpression(expression.Right)
-		return interpreter.testEqual(left, right)
-
-	case ast.OperationNotEqual:
-		left := interpreter.evalExpression(expression.Left)
-		right := interpreter.evalExpression(expression.Right)
-		return !interpreter.testEqual(left, right)
-
-	case ast.OperationOr:
-		// interpret the left-hand side
-		left := interpreter.evalExpression(expression.Left).(BoolValue)
-		// only interpret right-hand side if left-hand side is false
-		if left {
-			return left
-		}
-
-		// after interpreting the left-hand side,
-		// interpret the right-hand side
-		return interpreter.evalExpression(expression.Right).(BoolValue)
-
-	case ast.OperationAnd:
-		// interpret the left-hand side
-		left := interpreter.evalExpression(expression.Left).(BoolValue)
-		// only interpret right-hand side if left-hand side is true
-		if !left {
-			return left
-		}
-
-		// after interpreting the left-hand side,
-		// interpret the right-hand side
-		return interpreter.evalExpression(expression.Right).(BoolValue)
-
-	case ast.OperationNilCoalesce:
-		// interpret the left-hand side
-		left := interpreter.evalExpression(expression.Left)
-
-		// only evaluate right-hand side if left-hand side is nil
-		if some, ok := left.(*SomeValue); ok {
-			return some.Value
-		}
-
-		value := interpreter.evalExpression(expression.Right)
-
-		rightType := interpreter.Program.Elaboration.BinaryExpressionRightTypes[expression]
-		resultType := interpreter.Program.Elaboration.BinaryExpressionResultTypes[expression]
-
-		// NOTE: important to convert both any and optional
-		return interpreter.convertAndBox(value, rightType, resultType)
-	}
-
-	panic(&unsupportedOperation{
-		kind:      common.OperationKindBinary,
-		operation: expression.Operation,
-		Range:     ast.NewRangeFromPositioned(expression),
-	})
-}
-
-func (interpreter *Interpreter) testEqual(left, right Value) BoolValue {
-	left = interpreter.unbox(left)
-	right = interpreter.unbox(right)
-
-	// TODO: add support for arrays and dictionaries
-
-	switch left := left.(type) {
-	case NilValue:
-		_, ok := right.(NilValue)
-		return BoolValue(ok)
-
-	case EquatableValue:
-		// NOTE: might be NilValue
-		right, ok := right.(EquatableValue)
-		if !ok {
-			return false
-		}
-		return left.Equal(interpreter, right)
-
-	case *ArrayValue,
-		*DictionaryValue:
-		// TODO:
-		return false
-
-	default:
-		return false
-	}
-}
-
-func (interpreter *Interpreter) VisitUnaryExpression(expression *ast.UnaryExpression) ast.Repr {
-	value := interpreter.evalExpression(expression.Expression)
-
-	switch expression.Operation {
-	case ast.OperationNegate:
-		boolValue := value.(BoolValue)
-		return boolValue.Negate()
-
-	case ast.OperationMinus:
-		integerValue := value.(NumberValue)
-		return integerValue.Negate()
-
-	case ast.OperationMove:
-		return value
-	}
-
-	panic(&unsupportedOperation{
-		kind:      common.OperationKindUnary,
-		operation: expression.Operation,
-		Range:     ast.NewRangeFromPositioned(expression),
-	})
-}
-
-func (interpreter *Interpreter) VisitExpressionStatement(statement *ast.ExpressionStatement) ast.Repr {
-	result := interpreter.evalExpression(statement.Expression)
-	return Done{
-		Result: ExpressionStatementResult{result},
-	}
-}
-
-func (interpreter *Interpreter) VisitBoolExpression(expression *ast.BoolExpression) ast.Repr {
-	return BoolValue(expression.Value)
-}
-
-func (interpreter *Interpreter) VisitNilExpression(_ *ast.NilExpression) ast.Repr {
-	return NilValue{}
-}
-
-func (interpreter *Interpreter) VisitIntegerExpression(expression *ast.IntegerExpression) ast.Repr {
-	return IntValue{expression.Value}
-}
-
-func (interpreter *Interpreter) VisitFixedPointExpression(expression *ast.FixedPointExpression) ast.Repr {
-	// TODO: adjust once/if we support more fixed point types
-
-	value := fixedpoint.ConvertToFixedPointBigInt(
-		expression.Negative,
-		expression.UnsignedInteger,
-		expression.Fractional,
-		expression.Scale,
-		sema.Fix64Scale,
-	)
-
-	if expression.Negative {
-		return Fix64Value(value.Int64())
-	} else {
-		return UFix64Value(value.Uint64())
-	}
-}
-
-func (interpreter *Interpreter) VisitStringExpression(expression *ast.StringExpression) ast.Repr {
-	return NewStringValue(expression.Value)
-}
-
-func (interpreter *Interpreter) VisitArrayExpression(expression *ast.ArrayExpression) ast.Repr {
-	values := interpreter.visitExpressionsNonCopying(expression.Values)
-
-	argumentTypes := interpreter.Program.Elaboration.ArrayExpressionArgumentTypes[expression]
-	elementType := interpreter.Program.Elaboration.ArrayExpressionElementType[expression]
-
-	copies := make([]Value, len(values))
-	for i, argument := range values {
-		argumentType := argumentTypes[i]
-		copies[i] = interpreter.copyAndConvert(argument, argumentType, elementType)
-	}
-
-	return NewArrayValueUnownedNonCopying(copies...)
-}
-
-func (interpreter *Interpreter) VisitDictionaryExpression(expression *ast.DictionaryExpression) ast.Repr {
-	values := interpreter.visitEntries(expression.Entries)
-
-	entryTypes := interpreter.Program.Elaboration.DictionaryExpressionEntryTypes[expression]
-	dictionaryType := interpreter.Program.Elaboration.DictionaryExpressionType[expression]
-
-	dictionary := NewDictionaryValueUnownedNonCopying()
-	for i, dictionaryEntryValues := range values {
-		entryType := entryTypes[i]
-
-		key := interpreter.copyAndConvert(
-			dictionaryEntryValues.Key,
-			entryType.KeyType,
-			dictionaryType.KeyType,
-		)
-
-		value := interpreter.copyAndConvert(
-			dictionaryEntryValues.Value,
-			entryType.ValueType,
-			dictionaryType.ValueType,
-		)
-
-		// TODO: panic for duplicate keys?
-
-		// NOTE: important to convert in optional, as assignment to dictionary
-		// is always considered as an optional
-
-		locationRange := interpreter.locationRange(expression)
-		_ = dictionary.Insert(interpreter, locationRange, key, value)
-	}
-
-	return dictionary
-}
-
-func (interpreter *Interpreter) VisitMemberExpression(expression *ast.MemberExpression) ast.Repr {
-	result := interpreter.evalExpression(expression.Expression)
-	if expression.Optional {
-		switch typedResult := result.(type) {
-		case NilValue:
-			return typedResult
-
-		case *SomeValue:
-			result = typedResult.Value
-
-		default:
-			panic(errors.NewUnreachableError())
-		}
-	}
-
-	value := result.(Value)
-	locationRange := interpreter.locationRange(expression)
-	resultValue := interpreter.getMember(value, locationRange, expression.Identifier.Identifier)
-
-	// If the member access is optional chaining, only wrap the result value
-	// in an optional, if it is not already an optional value
-
-	if expression.Optional {
-		if _, ok := resultValue.(OptionalValue); !ok {
-			resultValue = NewSomeValueOwningNonCopying(resultValue)
-		}
-	}
-
-	return resultValue
-}
-
-func (interpreter *Interpreter) VisitIndexExpression(expression *ast.IndexExpression) ast.Repr {
-	typedResult := interpreter.evalExpression(expression.TargetExpression).(ValueIndexableValue)
-	indexingValue := interpreter.evalExpression(expression.IndexingExpression)
-	locationRange := interpreter.locationRange(expression)
-	return typedResult.Get(interpreter, locationRange, indexingValue)
-}
-
-func (interpreter *Interpreter) VisitConditionalExpression(expression *ast.ConditionalExpression) ast.Repr {
-	value := interpreter.evalExpression(expression.Test).(BoolValue)
-	if value {
-		return interpreter.evalExpression(expression.Then)
-	} else {
-		return interpreter.evalExpression(expression.Else)
-	}
-}
-
-func (interpreter *Interpreter) VisitInvocationExpression(invocationExpression *ast.InvocationExpression) ast.Repr {
-	// interpret the invoked expression
-	result := interpreter.evalExpression(invocationExpression.InvokedExpression)
-
-	// Handle optional chaining on member expression, if any:
-	// - If the member expression is nil, finish execution
-	// - If the member expression is some value, the wrapped value
-	//   is the function value that should be invoked
-
-	isOptionalChaining := false
-
-	if invokedMemberExpression, ok :=
-		invocationExpression.InvokedExpression.(*ast.MemberExpression); ok && invokedMemberExpression.Optional {
-
-		isOptionalChaining = true
-
-		switch typedResult := result.(type) {
-		case NilValue:
-			return typedResult
-
-		case *SomeValue:
-			result = typedResult.Value
-
-		default:
-			panic(errors.NewUnreachableError())
-		}
-	}
-
-	function := result.(FunctionValue)
-
-	// NOTE: evaluate all argument expressions in call-site scope, not in function body
-	argumentExpressions := make([]ast.Expression, len(invocationExpression.Arguments))
-	for i, argument := range invocationExpression.Arguments {
-		argumentExpressions[i] = argument.Expression
-	}
-
-	arguments := interpreter.visitExpressionsNonCopying(argumentExpressions)
-
-	typeParameterTypes :=
-		interpreter.Program.Elaboration.InvocationExpressionTypeArguments[invocationExpression]
-	argumentTypes :=
-		interpreter.Program.Elaboration.InvocationExpressionArgumentTypes[invocationExpression]
-	parameterTypes :=
-		interpreter.Program.Elaboration.InvocationExpressionParameterTypes[invocationExpression]
-
-	interpreter.reportFunctionInvocation(invocationExpression)
-
-	resultValue := interpreter.functionValueInvocationTrampoline(
-		function,
-		arguments,
-		argumentTypes,
-		parameterTypes,
-		typeParameterTypes,
-		ast.NewRangeFromPositioned(invocationExpression),
-	)
-
-	// If this is invocation is optional chaining, wrap the result
-	// as an optional, as the result is expected to be an optional
-
-	if isOptionalChaining {
-		resultValue = NewSomeValueOwningNonCopying(resultValue)
-	}
-
-	return resultValue
-}
-
-func (interpreter *Interpreter) InvokeFunctionValue(
-	function FunctionValue,
-	arguments []Value,
-	argumentTypes []sema.Type,
-	parameterTypes []sema.Type,
-	invocationRange ast.Range,
-) (value Value, err error) {
-	// recover internal panics and return them as an error
-	defer recoverErrors(func(internalErr error) {
-		err = internalErr
-	})
-
-	return interpreter.functionValueInvocationTrampoline(
-		function,
-		arguments,
-		argumentTypes,
-		parameterTypes,
-		nil,
-		invocationRange,
-	), nil
-}
-
-func (interpreter *Interpreter) functionValueInvocationTrampoline(
-	function FunctionValue,
-	arguments []Value,
-	argumentTypes []sema.Type,
-	parameterTypes []sema.Type,
-	typeParameterTypes *sema.TypeParameterTypeOrderedMap,
-	invocationRange ast.Range,
-) Value {
-
-	parameterTypeCount := len(parameterTypes)
-	argumentCopies := make([]Value, len(arguments))
-
-	for i, argument := range arguments {
-		argumentType := argumentTypes[i]
-		if i < parameterTypeCount {
-			parameterType := parameterTypes[i]
-			argumentCopies[i] = interpreter.copyAndConvert(argument, argumentType, parameterType)
-		} else {
-			argumentCopies[i] = argument.Copy()
-		}
-	}
-
-	// TODO: optimize: only potentially used by host-functions
-
-	locationRange := LocationRange{
-		Location: interpreter.Location,
-		Range:    invocationRange,
-	}
-
-	invocation := Invocation{
-		Arguments:          argumentCopies,
-		ArgumentTypes:      argumentTypes,
-		TypeParameterTypes: typeParameterTypes,
-		LocationRange:      locationRange,
-		Interpreter:        interpreter,
-	}
-
-	return function.Invoke(invocation)
-}
-
-func (interpreter *Interpreter) invokeInterpretedFunction(
-	function InterpretedFunctionValue,
-	invocation Invocation,
-) Value {
-
-	// Start a new activation record.
-	// Lexical scope: use the function declaration's activation record,
-	// not the current one (which would be dynamic scope)
-	interpreter.activations.PushNewWithParent(function.Activation)
-
-	// Make `self` available, if any
-	if invocation.Self != nil {
-		interpreter.declareVariable(sema.SelfIdentifier, invocation.Self)
-	}
-
-	return interpreter.invokeInterpretedFunctionActivated(function, invocation.Arguments)
-}
-
-// NOTE: assumes the function's activation (or an extension of it) is pushed!
-//
-func (interpreter *Interpreter) invokeInterpretedFunctionActivated(
-	function InterpretedFunctionValue,
-	arguments []Value,
-) Value {
-	defer interpreter.activations.Pop()
-
-	if function.ParameterList != nil {
-		interpreter.bindParameterArguments(function.ParameterList, arguments)
-	}
-
-	functionBlockTrampoline := interpreter.visitFunctionBody(
-		function.BeforeStatements,
-		function.PreConditions,
-		func() interface{} {
-			return interpreter.runAllStatements(interpreter.visitStatements(function.Statements))
-		},
-		function.PostConditions,
-		function.Type.ReturnTypeAnnotation.Type,
-	)
-
-	return interpreter.runAllStatements(functionBlockTrampoline).(Value)
-}
-
-// bindParameterArguments binds the argument values to the given parameters
-//
-func (interpreter *Interpreter) bindParameterArguments(
-	parameterList *ast.ParameterList,
-	arguments []Value,
-) {
-	for parameterIndex, parameter := range parameterList.Parameters {
-		argument := arguments[parameterIndex]
-		interpreter.declareVariable(parameter.Identifier.Identifier, argument)
-	}
-}
-
-func (interpreter *Interpreter) visitExpressionsNonCopying(expressions []ast.Expression) []Value {
-	values := make([]Value, 0, len(expressions))
-
-	for _, expression := range expressions {
-		value := interpreter.evalExpression(expression)
-		values = append(values, value)
-	}
-
-	return values
-}
-
-func (interpreter *Interpreter) visitEntries(entries []ast.DictionaryEntry) []DictionaryEntryValues {
-	values := make([]DictionaryEntryValues, 0, len(entries))
-
-	for _, entry := range entries {
-		key := interpreter.evalExpression(entry.Key)
-		value := interpreter.evalExpression(entry.Value)
-
-		values = append(
-			values,
-			DictionaryEntryValues{
-				Key:   key,
-				Value: value,
-			},
-		)
-	}
-
-	return values
-}
-
-func (interpreter *Interpreter) VisitFunctionExpression(expression *ast.FunctionExpression) ast.Repr {
-
-	// lexical scope: variables in functions are bound to what is visible at declaration time
-	lexicalScope := interpreter.activations.CurrentOrNew()
-
-	functionType := interpreter.Program.Elaboration.FunctionExpressionFunctionType[expression]
-
-	var preConditions ast.Conditions
-	if expression.FunctionBlock.PreConditions != nil {
-		preConditions = *expression.FunctionBlock.PreConditions
-	}
-
-	var beforeStatements []ast.Statement
-	var rewrittenPostConditions ast.Conditions
-
-	if expression.FunctionBlock.PostConditions != nil {
-		postConditionsRewrite :=
-			interpreter.Program.Elaboration.PostConditionsRewrite[expression.FunctionBlock.PostConditions]
-
-		rewrittenPostConditions = postConditionsRewrite.RewrittenPostConditions
-		beforeStatements = postConditionsRewrite.BeforeStatements
-	}
-
-	statements := expression.FunctionBlock.Block.Statements
-
-	return InterpretedFunctionValue{
-		Interpreter:      interpreter,
-		ParameterList:    expression.ParameterList,
-		Type:             functionType,
-		Activation:       lexicalScope,
-		BeforeStatements: beforeStatements,
-		PreConditions:    preConditions,
-		Statements:       statements,
-		PostConditions:   rewrittenPostConditions,
-	}
 }
 
 // NOTE: only called for top-level composite declarations
@@ -3099,300 +2100,6 @@ func (interpreter *Interpreter) NewSubInterpreter(
 			options...,
 		)...,
 	)
-}
-
-func (interpreter *Interpreter) VisitPragmaDeclaration(_ *ast.PragmaDeclaration) ast.Repr {
-	return Done{}
-}
-
-func (interpreter *Interpreter) VisitImportDeclaration(declaration *ast.ImportDeclaration) ast.Repr {
-
-	resolvedLocations := interpreter.Program.Elaboration.ImportDeclarationsResolvedLocations[declaration]
-
-	for _, resolvedLocation := range resolvedLocations {
-		interpreter.importResolvedLocation(resolvedLocation)
-	}
-
-	return Done{}
-}
-
-func (interpreter *Interpreter) importResolvedLocation(resolvedLocation sema.ResolvedLocation) {
-
-	subInterpreter := interpreter.ensureLoaded(
-		resolvedLocation.Location,
-		func() Import {
-			return interpreter.importLocationHandler(interpreter, resolvedLocation.Location)
-		},
-	)
-
-	// determine which identifiers are imported /
-	// which variables need to be declared
-
-	var variables map[string]*Variable
-	identifierLength := len(resolvedLocation.Identifiers)
-	if identifierLength > 0 {
-		variables = make(map[string]*Variable, identifierLength)
-		for _, identifier := range resolvedLocation.Identifiers {
-			variables[identifier.Identifier] =
-				subInterpreter.Globals[identifier.Identifier]
-		}
-	} else {
-		variables = subInterpreter.Globals
-	}
-
-	// Gather all variable names and sort them lexicographically
-
-	var names []string
-
-	for name := range variables { //nolint:maprangecheck
-		names = append(names, name)
-	}
-
-	// Set variables for all imported values in lexicographic order
-
-	sort.Strings(names)
-
-	for _, name := range names {
-		variable := variables[name]
-
-		// don't import predeclared values
-		if subInterpreter.Program != nil {
-			if _, ok := subInterpreter.Program.Elaboration.EffectivePredeclaredValues[name]; ok {
-				continue
-			}
-		}
-
-		// don't import base values
-		if sema.BaseValueActivation.Find(name) != nil {
-			continue
-		}
-
-		interpreter.setVariable(name, variable)
-		interpreter.Globals[name] = variable
-	}
-}
-
-func (interpreter *Interpreter) VisitTransactionDeclaration(declaration *ast.TransactionDeclaration) ast.Repr {
-	interpreter.declareTransactionEntryPoint(declaration)
-
-	// NOTE: no result, so it does *not* act like a return-statement
-	return Done{}
-}
-
-func (interpreter *Interpreter) declareTransactionEntryPoint(declaration *ast.TransactionDeclaration) {
-	transactionType := interpreter.Program.Elaboration.TransactionDeclarationTypes[declaration]
-
-	lexicalScope := interpreter.activations.CurrentOrNew()
-
-	var prepareFunction *ast.FunctionDeclaration
-	var prepareFunctionType *sema.FunctionType
-	if declaration.Prepare != nil {
-		prepareFunction = declaration.Prepare.FunctionDeclaration
-		prepareFunctionType = transactionType.PrepareFunctionType().InvocationFunctionType()
-	}
-
-	var executeFunction *ast.FunctionDeclaration
-	var executeFunctionType *sema.FunctionType
-	if declaration.Execute != nil {
-		executeFunction = declaration.Execute.FunctionDeclaration
-		executeFunctionType = transactionType.ExecuteFunctionType().InvocationFunctionType()
-	}
-
-	postConditionsRewrite :=
-		interpreter.Program.Elaboration.PostConditionsRewrite[declaration.PostConditions]
-
-	self := &CompositeValue{
-		Location: interpreter.Location,
-		Fields:   NewStringValueOrderedMap(),
-		modified: true,
-	}
-
-	transactionFunction := NewHostFunctionValue(
-		func(invocation Invocation) Value {
-			interpreter.activations.PushNewWithParent(lexicalScope)
-
-			invocation.Self = self
-			interpreter.declareVariable(sema.SelfIdentifier, self)
-
-			if declaration.ParameterList != nil {
-				// If the transaction has a parameter list of N parameters,
-				// bind the first N arguments of the invocation to the transaction parameters,
-				// then leave the remaining arguments for the prepare function
-
-				transactionParameterCount := len(declaration.ParameterList.Parameters)
-
-				transactionArguments := invocation.Arguments[:transactionParameterCount]
-				prepareArguments := invocation.Arguments[transactionParameterCount:]
-
-				interpreter.bindParameterArguments(declaration.ParameterList, transactionArguments)
-				invocation.Arguments = prepareArguments
-			}
-
-			// NOTE: get current scope instead of using `lexicalScope`,
-			// because current scope has `self` declared
-			transactionScope := interpreter.activations.CurrentOrNew()
-
-			if prepareFunction != nil {
-				prepare := interpreter.functionDeclarationValue(
-					prepareFunction,
-					prepareFunctionType,
-					transactionScope,
-				)
-
-				prepare.Invoke(invocation)
-			}
-
-			var body func() interface{}
-			if executeFunction != nil {
-				execute := interpreter.functionDeclarationValue(
-					executeFunction,
-					executeFunctionType,
-					transactionScope,
-				)
-
-				invocationWithoutArguments := invocation
-				invocationWithoutArguments.Arguments = nil
-
-				body = func() interface{} {
-					return execute.Invoke(invocationWithoutArguments)
-				}
-			}
-
-			var preConditions ast.Conditions
-			if declaration.PreConditions != nil {
-				preConditions = *declaration.PreConditions
-			}
-
-			tram := interpreter.visitFunctionBody(
-				postConditionsRewrite.BeforeStatements,
-				preConditions,
-				body,
-				postConditionsRewrite.RewrittenPostConditions,
-				sema.VoidType,
-			)
-
-			return interpreter.runAllStatements(tram).(Value)
-		})
-
-	interpreter.Transactions = append(interpreter.Transactions, &transactionFunction)
-}
-
-func (interpreter *Interpreter) VisitEmitStatement(statement *ast.EmitStatement) ast.Repr {
-	event := interpreter.evalExpression(statement.InvocationExpression).(*CompositeValue)
-
-	eventType := interpreter.Program.Elaboration.EmitStatementEventTypes[statement]
-
-	if interpreter.onEventEmitted == nil {
-		panic(EventEmissionUnavailableError{
-			LocationRange: interpreter.locationRange(statement),
-		})
-	}
-
-	err := interpreter.onEventEmitted(interpreter, event, eventType)
-	if err != nil {
-		panic(err)
-	}
-
-	// NOTE: no result, so it does *not* act like a return-statement
-	return Done{}
-}
-
-func (interpreter *Interpreter) VisitCastingExpression(expression *ast.CastingExpression) ast.Repr {
-	value := interpreter.evalExpression(expression.Expression)
-
-	expectedType := interpreter.Program.Elaboration.CastingTargetTypes[expression]
-
-	switch expression.Operation {
-	case ast.OperationFailableCast, ast.OperationForceCast:
-		dynamicType := value.DynamicType(interpreter)
-		isSubType := IsSubType(dynamicType, expectedType)
-
-		switch expression.Operation {
-		case ast.OperationFailableCast:
-			if !isSubType {
-				return NilValue{}
-			}
-
-			return NewSomeValueOwningNonCopying(value)
-
-		case ast.OperationForceCast:
-			if !isSubType {
-				panic(
-					TypeMismatchError{
-						ExpectedType:  expectedType,
-						LocationRange: interpreter.locationRange(expression.Expression),
-					},
-				)
-			}
-
-			return value
-
-		default:
-			panic(errors.NewUnreachableError())
-		}
-
-	case ast.OperationCast:
-		staticValueType := interpreter.Program.Elaboration.CastingStaticValueTypes[expression]
-		return interpreter.convertAndBox(value, staticValueType, expectedType)
-
-	default:
-		panic(errors.NewUnreachableError())
-	}
-}
-
-func (interpreter *Interpreter) VisitCreateExpression(expression *ast.CreateExpression) ast.Repr {
-	return interpreter.evalExpression(expression.InvocationExpression)
-}
-
-func (interpreter *Interpreter) VisitDestroyExpression(expression *ast.DestroyExpression) ast.Repr {
-	value := interpreter.evalExpression(expression.Expression)
-
-	// TODO: optimize: only potentially used by host-functions
-
-	locationRange := interpreter.locationRange(expression)
-
-	// TODO: change Destroy to return Value
-	return interpreter.runAllStatements(value.(DestroyableValue).Destroy(interpreter, locationRange))
-}
-
-func (interpreter *Interpreter) VisitReferenceExpression(referenceExpression *ast.ReferenceExpression) ast.Repr {
-
-	authorized := referenceExpression.Type.(*ast.ReferenceType).Authorized
-
-	result := interpreter.evalExpression(referenceExpression.Expression)
-
-	return &EphemeralReferenceValue{
-		Authorized: authorized,
-		Value:      result,
-	}
-}
-
-func (interpreter *Interpreter) VisitForceExpression(expression *ast.ForceExpression) ast.Repr {
-	result := interpreter.evalExpression(expression.Expression)
-
-	switch result := result.(type) {
-	case *SomeValue:
-		return result.Value
-
-	case NilValue:
-		panic(
-			ForceNilError{
-				LocationRange: interpreter.locationRange(expression.Expression),
-			},
-		)
-
-	default:
-		return result
-	}
-}
-
-func (interpreter *Interpreter) VisitPathExpression(expression *ast.PathExpression) ast.Repr {
-	domain := common.PathDomainFromIdentifier(expression.Domain.Identifier)
-
-	return PathValue{
-		Domain:     domain,
-		Identifier: expression.Identifier.Identifier,
-	}
 }
 
 func (interpreter *Interpreter) storedValueExists(storageAddress common.Address, key string) bool {
